+++ packages/sandbox/lib/router/guards/auth_guard.dart
@@ -24,7 +24,8 @@ class AuthGuard extends AutoRouteGuard {
         resolver.next();
       case AuthenticationStatus.unauthenticated:
       case AuthenticationStatus.authenticated:
-        resolver.redirect(const DashboardRouter());
+        router.push(const DashboardRouter());
+        resolver.next(false);
     }
   }
 }